部分CI&prompt模板&check.py prompt格式检查脚本

+++ .ci/prompt_to_json_for_CI.py
@@ -4,7 +4,7 @@ import re
 import os
 '''
 # @author  : Shiqiding
-# @description: 本脚本支持将.md写的规定prompt格式转换为规定格式的json,仅支持单文本转换为CI pipeline准备
+# @description: 本脚本支持将批量将.md写的规定prompt格式转换为规定格式的json
 # @version : V2.0
 
 promptpath为.md格式的prompt路径，具体格式参考仓库里的prompt格式部分
@@ -35,9 +35,9 @@ input:用户输入
 output:对应输出
 
 '''
-
-promptpath=r'C:\Users\yhd\PycharmProjects\Tianji\test\gpt_prompt\02-Hospitality\02-Hospitality-职场如何发出邀请_上级、客户、友商、同事等多场景.md'
-
+folder_path = os.environ.get('folder_path')
+#folder_path = r"C:\Users\yhd\PycharmProjects\Tianji\test\prompt\yiyan_prompt"  # 替换成您的文件夹路径
+output_path=os.environ.get('output_path')
 
 def md_file_to_json_with_examples(file_path,id,heading):
     """
@@ -61,7 +61,7 @@ def md_file_to_json_with_examples(file_path,id,heading):
 
         for block in blocks:
             test_system_part = block.split('#### Q：')[0].strip()
-            test_system_part = re.sub(r'#.*', '', test_system_part)
+
             qa_pairs = re.findall(r'#### Q：(.*?)#### A：(.*?)(?=#### Q：|$)', block, re.DOTALL)
             if (qa_pairs == []):
                 qa_pairs = re.findall(r'#### Q:(.*?)#### A:(.*?)(?=#### Q:|$)', block, re.DOTALL)
@@ -69,9 +69,7 @@ def md_file_to_json_with_examples(file_path,id,heading):
             examples = []
             for qa_pair in qa_pairs:
                 input_text = qa_pair[0].strip()
-                input_text = re.sub(r'#.*', '', input_text)
                 output_text = qa_pair[1].strip()
-                output_text = re.sub(r'#.*', '', output_text)
 
                 example_obj = {
                     "id":id,
@@ -175,29 +173,36 @@ def find_first_heading(md_file_path):
 
 
 if __name__ == '__main__':
-    filepath =replace_english_colons_with_chinese(promptpath)
-    heading = find_first_heading(filepath)
-    print("此文档的heading使用的是 "+heading)
-    filename=os.path.basename(filepath)
-    id =int(filename[:2])
-    print("处理文档为 " + filename+" 该文档属于第"+str(id)+"大类")
-    json_output = md_file_to_json_with_examples(filepath,id=id,heading=heading)
-    input_dir, input_file = os.path.split(promptpath)
-    input_file_base, _ = os.path.splitext(input_file)
 
-    output_path = r"/tianji/prompt"
+    for foldername, subfolders, filenames in os.walk(folder_path):
+        for filename in filenames:
+            if filename.endswith(".md"):
+                if filename == "README.md":
+                    continue
+                promptpath=os.path.join(foldername, filename)
+                filepath =replace_english_colons_with_chinese(promptpath)
+                print(filepath)
+                heading = find_first_heading(filepath)
+                print("此文档的heading使用的是 "+heading)
+                filename=os.path.basename(filepath)
+                id =int(filename[:2])
+                print("处理文档为 " + filename+" 该文档属于第"+str(id)+"大类")
+                json_output = md_file_to_json_with_examples(filepath,id=id,heading=heading)
+                input_dir, input_file = os.path.split(promptpath)
+                input_file_base, _ = os.path.splitext(input_file)
 
-    # 使用正则表达式提取所需路径
-    match = re.search(r'\\test(.*)\\[^\\]+$', promptpath)
-    if match:
-        # 提取的路径
-        extracted_path = match.group(1)
-        # 构造最终路径
-        json_file_output_path = output_path + extracted_path + "\\"
-    else:
-        json_file_output_path = "无法匹配路径"
-    json_file_output = os.path.join(json_file_output_path, input_file_base + ".json")
-    with open(json_file_output, 'w', encoding='utf-8') as file:
-        json.dump(json.loads(json_output), file, ensure_ascii=False, indent=4)
-    print(json_output)
+                #output_path = r"C:\Users\yhd\PycharmProjects\Tianji\tianji\prompt"
 
+                # 使用正则表达式提取所需路径
+                match = re.search(r'/prompt(.*)/[^/]+$', promptpath)
+                if match:
+                    # 提取的路径
+                    extracted_path = match.group(1)
+                    # 构造最终路径
+                    json_file_output_path = output_path + extracted_path + "/"
+                else:
+                    json_file_output_path = "无法匹配路径"
+                json_file_output = os.path.join(json_file_output_path, input_file_base + ".json")
+                with open(json_file_output, 'w', encoding='utf-8') as file:
+                    json.dump(json.loads(json_output), file, ensure_ascii=False, indent=4)
+                print(json_output)
